BDD/requirements.txt
# ============================================================================
# PYTHON PACKAGE REQUIREMENTS:
# ============================================================================
# REQUIRES: Python >= 3.8.7
# DESCRIPTION:
#    pip install -r <THIS_FILE>
#
# SPECIFIC DEPENDENCIES:
#   behave  -- BDD test framework, cucumber-clone for Python.
#   selenium -- To drive the browser tests
#   PyHamcrest -- For assertions: assert_that(expected, equal_to(actual))
#   PyYAML -- For working with yaml files
#
# MAYBE:
#   nose:       for assertions: eq_(x, y)
#   pytest:     for assertions
# ============================================================================

behave >= 1.2.6
behave-html-formatter
PyHamcrest